requirements.txt
python-dotenv
crewai
crewai_tools
fastapi
uvicorn
langchain_community
pypdf
python-multipart
pip install langchain-ollama